@@ -8,8 +8,8 @@
     "@material-ui/icons": "^4.11.3",
     "@material-ui/lab": "^4.0.0-alpha.61",
     "@material-ui/pickers": "^3.3.10",
-    "@tauri-apps/api": "^1.0.0-rc.5",
-    "@tauri-apps/cli": "^1.0.0-rc.11",
+    "@tauri-apps/api": "^1.0.0-rc.6",
+    "@tauri-apps/cli": "^1.0.0-rc.12",
     "ajv": "^8.11.0",
     "date-fns": "^2.28.0",
     "lodash": "^4.17.21",
@@ -23,16 +23,16 @@
     "react-helmet": "^6.1.0",
     "react-hook-form": "^7.22.5",
     "react-number-format": "^4.9.3",
-    "react-redux": "^8.0.1",
-    "react-router": "^5.3.1",
-    "react-router-dom": "^5.3.3",
+    "react-redux": "^8.0.2",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "^4.0.3",
     "redux": "^4.2.0",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
-    "rooks": "^5.11.0",
+    "rooks": "^5.11.2",
     "ts-json-schema-generator": "^1.0.0",
-    "typescript": "^4.6.4"
+    "typescript": "^4.7.2"
   },
   "scripts": {
     "start": "node ./notice.js && ts-node ./scripts/env.ts && ts-node ./scripts/setup.ts && concurrently npm:dev",
@@ -73,17 +73,17 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.1",
     "@types/lodash": "^4.14.182",
-    "@types/node": "^17.0.33",
+    "@types/node": "^17.0.35",
     "@types/papaparse": "^5.3.2",
     "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.4",
+    "@types/react-dom": "^18.0.5",
     "@types/react-helmet": "^6.1.5",
     "@types/react-redux": "^7.1.24",
     "@types/react-router-dom": "^5.3.3",
     "@types/redux-logger": "^3.0.9",
-    "@typescript-eslint/eslint-plugin": "^5.25.0",
-    "@typescript-eslint/parser": "^5.25.0",
-    "concurrently": "^7.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^5.26.0",
+    "concurrently": "^7.2.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
@@ -92,10 +92,10 @@
     "eslint-plugin-react": "^7.30.0",
     "fast-check": "^2.25.0",
     "husky": "^8.0.1",
-    "lint-staged": "^12.4.1",
+    "lint-staged": "^12.4.2",
     "prettier": "2.6.2",
     "react-router-last-location": "^2.0.1",
-    "ts-node": "^10.7.0",
+    "ts-node": "^10.8.0",
     "typescript-eslint": "0.0.1-alpha.0"
   },
   "lint-staged": {
