@@ -1,15 +1,12 @@
-import { useMemo, PropsWithChildren, ReactElement, ReactNode } from 'react'
+import { PropsWithChildren, ReactElement } from 'react'
 import {
   createStyles,
   makeStyles,
-  unstable_createMuiStrictModeTheme as createMuiTheme,
-  useMediaQuery,
   CssBaseline,
   Grid,
-  Theme,
-  ThemeProvider
+  Theme
 } from '@material-ui/core'
-import { Switch, Route, Redirect, useLocation } from 'react-router-dom'
+import { Routes, Route, Navigate } from 'react-router-dom'
 import { isMobileOnly as isMobile } from 'react-device-detect'
 import { PagerProvider } from './pager'
 import { StateLoader } from './debug'
@@ -19,7 +16,6 @@ import ScrollTop from './ScrollTop'
 import Menu, { backPages, drawerSectionsForYear } from './Menu'
 import { Section, SectionItem } from './ResponsiveDrawer'
 
-import { useFocus } from 'ustaxes/hooks/Focus'
 import Urls from 'ustaxes/data/urls'
 import DataPropagator from './DataPropagator'
 import YearStatusBar from './YearStatusBar'
@@ -57,41 +53,8 @@ const useStyles = makeStyles<Theme, Props>((theme: Theme) =>
 )
 
 export default function Main(): ReactElement {
-  const [ref] = useFocus()
-
   const activeYear: TaxYear = useSelector(
     (state: YearsTaxesState) => state.activeYear
-  )
-
-  const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)')
-  const isStartPage = useLocation().pathname === '/start'
-  const theme = useMemo(
-    () =>
-      createMuiTheme({
-        palette: {
-          type: prefersDarkMode ? 'dark' : 'light',
-          secondary: prefersDarkMode
-            ? {
-                light: '#4f5b62',
-                main: '#d5d5d5',
-                dark: '#000a12',
-                contrastText: '#ffffff'
-              }
-            : {
-                light: '#4f5b62',
-                main: '#263238',
-                dark: '#000a12',
-                contrastText: '#ffffff'
-              },
-          primary: {
-            light: '#66ffa6',
-            main: '#00e676',
-            dark: '#00b248',
-            contrastText: '#000000'
-          }
-        }
-      }),
-    [prefersDarkMode]
   )
 
   const classes = useStyles({ isMobile })
@@ -102,57 +65,70 @@ export default function Main(): ReactElement {
 
   const allItems: SectionItem[] = [...steps, ...backPages]
 
-  const Layout = ({ children }: PropsWithChildren<{ children: ReactNode }>) => (
-    <Grid
-      ref={ref}
-      component="main"
-      tabIndex={-1}
-      container
-      justifyContent="center"
-      direction="row"
-    >
-      <Grid item sm={12} md={8} lg={6}>
-        {!isStartPage && (
+  const Layout = ({
+    showMenu = true,
+    children
+  }: PropsWithChildren<{ showMenu?: boolean }>) => (
+    <>
+      {showMenu ? <Menu /> : undefined}
+      <Grid
+        component="main"
+        tabIndex={-1}
+        container
+        justifyContent="center"
+        direction="row"
+      >
+        <Grid item sm={12} md={8} lg={6}>
+          {showMenu ? (
+            <Grid item className={classes.content}>
+              {' '}
+              <YearStatusBar />
+            </Grid>
+          ) : undefined}
           <Grid item className={classes.content}>
-            {' '}
-            <YearStatusBar />
+            {isMobile && showMenu ? (
+              <div className={classes.toolbar} />
+            ) : undefined}
+            {children}
           </Grid>
-        )}
-        <Grid item className={classes.content}>
-          {children}
         </Grid>
       </Grid>
-    </Grid>
+    </>
   )
 
   return (
-    <ThemeProvider theme={theme}>
+    <>
       <CssBaseline />
       <SkipToLinks />
-      {isMobile && !isStartPage && <div className={classes.toolbar} />}
       <div className={classes.container}>
         <StateLoader />
         <PagerProvider pages={steps}>
-          <Switch>
-            <Redirect path="/" to={Urls.default} exact />
+          <Routes>
+            <Route path="/" element={<Navigate to={Urls.default} />} />
             {allItems.map((item) => (
-              <Route key={item.title} exact path={item.url}>
-                {!isStartPage && <Menu />}
-                <Layout>
-                  {!isStartPage && <DataPropagator />}
-                  {item.element}
-                </Layout>
-              </Route>
+              <Route
+                key={item.title}
+                path={item.url}
+                element={
+                  <Layout showMenu={!item.url.includes('start')}>
+                    <DataPropagator />
+                    {item.element}
+                  </Layout>
+                }
+              />
             ))}
-            <Route>
-              <Layout>
-                <NoMatchPage />
-              </Layout>
-            </Route>
-          </Switch>
+            <Route
+              path="*"
+              element={
+                <Layout>
+                  <NoMatchPage />
+                </Layout>
+              }
+            />
+          </Routes>
         </PagerProvider>
         {!isMobile && <ScrollTop />}
       </div>
-    </ThemeProvider>
+    </>
   )
 }
