@@ -1,11 +1,46 @@
-import { ReactElement } from 'react'
+import { ReactElement, useMemo } from 'react'
 import Main from './components/Main'
 import './App.css'
+import { createTheme, ThemeProvider, useMediaQuery } from '@material-ui/core'
 
-const App = (): ReactElement => (
-  <div className="App">
-    <Main />
-  </div>
-)
+const App = (): ReactElement => {
+  const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)')
+
+  const theme = useMemo(
+    () =>
+      createTheme({
+        palette: {
+          secondary: prefersDarkMode
+            ? {
+                light: '#4f5b62',
+                main: '#d5d5d5',
+                dark: '#000a12',
+                contrastText: '#ffffff'
+              }
+            : {
+                light: '#4f5b62',
+                main: '#263238',
+                dark: '#000a12',
+                contrastText: '#ffffff'
+              },
+          primary: {
+            light: '#66ffa6',
+            main: '#00e676',
+            dark: '#00b248',
+            contrastText: '#000000'
+          }
+        }
+      }),
+    [prefersDarkMode]
+  )
+
+  return (
+    <div className="App">
+      <ThemeProvider theme={theme}>
+        <Main />
+      </ThemeProvider>
+    </div>
+  )
+}
 
 export default App
